Issue #1167 Implement Group Member Manage, Group manager fixes
 * Adding and removing members now works in the Member manager popup
 * Implement an intelligent research input for adding users
 * Add the plugin confirm.js
 * Add a confirm popup for the group deletion
 * Implement the "Set as group for new users" action for groups

+++ admin/themes/default/js/group_list.js
@@ -1,3 +1,5 @@
+const DELAY_FEEDBACK = 3000;
+
 /*-------
 Group Popup
 -------*/
@@ -17,7 +19,7 @@ $(".CloseGroupPopup").click(function () {
 $("#form-btn").click(function () {
   $("#cancel").show();
   $("#addUserLabel").show();
-  $(".input-user-name").show();
+  $(".UserSearch").show();
   $("#UserSubmit").show();
   $("#form-btn").hide();
   $(".groups .list_user").css("max-height", "100px");
@@ -26,7 +28,7 @@ $("#form-btn").click(function () {
 $("#cancel").click(function () {
   $("#cancel").hide();
   $("#addUserLabel").hide();
-  $(".input-user-name").hide();
+  $(".UserSearch").hide();
   $("#UserSubmit").hide();
   $("#form-btn").show();
   $(".groups .list_user").css("max-height", "200px");
@@ -86,7 +88,8 @@ jQuery(document).ready(function () {
           newgroup.find(".Group-checkbox label").attr("for", "Group-Checkbox-selection-" + id);
           newgroup.find(".Group-checkbox input").attr("id", "Group-Checkbox-selection-" + id);
           newgroup.find(".input-edit-group-name").attr("placeholder", name);
-          newgroup.find(".group_number_users").html("0 " + member_default);
+          newgroup.find(".group_number_users").html("0 " + str_member_default);
+          newgroup.find(".group_name-editable").html(name);
           hideAddGroupForm();
           //Setup the icon color
           var colors = [["#ffa744", "#ffe9cf"],["#896af3", "#e0daf4"], ["#6ece5e","#d6ffcf"],["#2883c3","#cfebff"]];
@@ -97,13 +100,13 @@ jQuery(document).ready(function () {
           
           //Place group in first Place 
           newgroup.appendTo(".groups");
-          newgroup.find(".groupMessage").html(group_created);
+          newgroup.find(".groupMessage").html(str_group_created);
           newgroup.find(".groupMessage").fadeIn();
-          newgroup.find(".groupMessage").delay(4000).fadeOut();
+          newgroup.find(".groupMessage").delay(DELAY_FEEDBACK).fadeOut();
         } else {
-          $("#showAddGroup .groupError").html(name_taken);
+          $("#showAddGroup .groupError").html(str_name_taken);
           $("#showAddGroup .groupError").fadeIn();
-          $("#showAddGroup .groupError").delay(3000).fadeOut();
+          $("#showAddGroup .groupError").delay(DELAY_FEEDBACK).fadeOut();
         }
       },
       error: function (err) {
@@ -112,24 +115,6 @@ jQuery(document).ready(function () {
     });
   });
 });
-
-/*-------
- User List group pop-in
- -------*/
-
-var modalUL = document.getElementById("UserList");
-
-var btn = document.getElementById("UserListTrigger");
-
-var close = document.getElementsByClassName("CloseUserList")[0];
-
-btn.onclick = function () {
-  modalUL.style.visibility = "visible";
-};
-
-close.onclick = function () {
-  modalUL.style.visibility = "hidden";
-};
 
 /*-------
  SETUP JS ON GROUP BOX
@@ -168,6 +153,8 @@ var setupGroupBox = function (groupBox) {
           .trigger("change");
       });
       updateSelectionPanel();
+      option = $('<option value="'+id+'">'+groupBox.find("#group_name").html()+'</option>')
+      option.appendTo("#MergeOptionsChoices");
     } else {
       groupBox.removeClass("OrangeBackground");
       groupBox.find(".icon-users-1").removeClass("OrangeIcon");
@@ -178,6 +165,11 @@ var setupGroupBox = function (groupBox) {
         }
       });
       updateSelectionPanel();
+      $("#MergeOptionsChoices option").each(function () {
+        if ($(this).attr("value") == id) {
+          $(this).remove();
+        }
+      });
     }
   });
 
@@ -188,25 +180,53 @@ var setupGroupBox = function (groupBox) {
 
   /* Set the delete action */
   groupBox.find("#GroupDelete").on("click", function () {
-    deleteGroup(id);
+    $.confirm({
+      title: str_delete+' '+groupBox.find("#group_name").html(),
+      content: str_are_you_sure,
+      boxWidth: '30%',
+      useBootstrap: false,
+      type: 'red',
+      typeAnimated: true,
+      buttons: {
+          confirm: {
+            text: str_yes_delete_confirmation,
+            btnClass: 'btn-red',
+            action: function () {
+              deleteGroup(id);
+            }
+          },
+          cancel: {
+            text: str_no_delete_confirmation
+          }
+      }
+    });
   });
 
   /* Set the rename action */
   groupBox.find(".Group-name .icon-pencil").on("click", function () {
-    groupBox.find(".Group-name-container form").show();
+    console.log("clic");
+    groupBox.find(".group-rename").css("display", "flex");
     groupBox.find(".Group-name-container .icon-pencil").hide();
-    groupBox.find(".Group-name-container .icon-ok").show();
     groupBox.find(".Group-name-container p").css("opacity", 0)
   });
 
   groupBox.find(".Group-name-container .icon-ok").on("click", function () {
-    renameGroup(id, groupBox.find(".group_name-editable").val())
+    if (groupBox.find(".group_name-editable").val() != groupBox.find("#group_name").html())
+      renameGroup(id, groupBox.find(".group_name-editable").val())
   });
 
   groupBox.find(".Group-name-container form").on("submit", function (e) {
     e.preventDefault();
-    renameGroup(id, groupBox.find(".group_name-editable").val())
+    if (groupBox.find(".group_name-editable").val() != groupBox.find("#group_name").html())
+      renameGroup(id, groupBox.find(".group_name-editable").val())
   });
+
+  groupBox.find(".group-rename .icon-cancel").on('click', function() {
+    groupBox.find(".group-rename").hide();
+    groupBox.find(".Group-name-container .icon-pencil").removeAttr("style");
+    groupBox.find(".Group-name-container p").css("opacity", 1);
+    groupBox.find(".group_name-editable").val(groupBox.find(".Group-name-container p").html());
+  })
 
   /* Hide group options and rename field on click on the screen */
 
@@ -214,14 +234,18 @@ var setupGroupBox = function (groupBox) {
     if ($(e.target).closest("#group-"+id+" #GroupOptions").length === 0) {
       groupBox.find(".group-dropdown-options #GroupOptions").hide();
     }
-    if ($(e.target).closest("#group-"+id+" .Group-name").length === 0) {
-      groupBox.find(".Group-name-container form").hide();
-      groupBox.find(".Group-name-container .icon-pencil").removeAttr("style");
-      groupBox.find(".Group-name-container .icon-ok").hide();
-      groupBox.find(".Group-name-container p").css("opacity", 1);
-      groupBox.find(".group_name-editable").val(groupBox.find(".Group-name-container p").html());
-    }
   });
+
+  /* Setup the default action */
+  if (groupBox.data("default") == 1) {
+    setupDefaultActions(id, true);
+  } else if (groupBox.data("default") == 0) {
+    setupDefaultActions(id, false);
+  }
+
+  groupBox.find(".manage-users").on("click", function(){openUserManager(id)});
+  
+
 };
 
 /* Group Ajax Functions */
@@ -233,6 +257,12 @@ var deleteGroup = function (id) {
     success: function (raw_data) {
       data = jQuery.parseJSON(raw_data);
       if (data.stat === "ok") {
+        $.alert({
+          title: str_group_deleted,
+          content: "",
+          boxWidth: '20%',
+          useBootstrap: false
+        });
         $("#group-" + id).remove();
       }
     },
@@ -252,27 +282,69 @@ var renameGroup = function(id, newName) {
       if (data.stat === "ok") {
         newName = data.result.groups[0].name;
         //Display message
-        $("#group-" + id).find(".groupMessage").html(renaming_done);
+        $("#group-" + id).find(".groupMessage").html(str_renaming_done);
         $("#group-" + id).find(".groupMessage").fadeIn();
-        $("#group-" + id).find(".groupMessage").delay(3000).fadeOut();
+        $("#group-" + id).find(".groupMessage").delay(DELAY_FEEDBACK).fadeOut();
         $("#group-" + id).find("#group_name").html(newName);
 
         //Hide editable field
-        $("#group-" + id).find(".Group-name-container form").hide();
-        $("#group-" + id).find(".Group-name-container span").show();
-        $("#group-" + id).find(".Group-name-container .icon-ok").hide();
-        $("#group-" + id).find(".Group-name-container p").css("opacity", 1)
+        $("#group-" + id).find(".group-rename").hide();
+        $("#group-" + id).find(".Group-name-container .icon-pencil").removeAttr("style");
+        $("#group-" + id).find(".Group-name-container p").css("opacity", 1);
       } else {
         //Display error message
-        $("#group-" + id).find(".groupError").html(name_taken);
+        $("#group-" + id).find(".groupError").html(str_name_taken);
         $("#group-" + id).find(".groupError").fadeIn();
-        $("#group-" + id).find(".groupError").delay(3000).fadeOut();
+        $("#group-" + id).find(".groupError").delay(DELAY_FEEDBACK).fadeOut();
       }
     },
     error: function (err) {
       console.log(err);
     },
   });
+}
+
+var setDefaultGroup = function (id, is_default) {
+  jQuery.ajax({
+    url: "ws.php?format=json&method=pwg.groups.setInfo",
+    type: "POST",
+    data: "group_id=" + id + "&pwg_token=" + pwg_token + "&is_default="+is_default,
+    success: function (raw_data) {
+      data = jQuery.parseJSON(raw_data);
+      if (data.stat === "ok") {
+        if (is_default) {
+          setupDefaultActions(id,true)
+          $("#group-" + id).find(".groupMessage").html(str_set_default_state);
+        } else {
+          setupDefaultActions(id,false)
+          $("#group-" + id).find(".groupMessage").html(str_unset_default_state);
+        }
+        $("#group-" + id).find(".groupMessage").fadeIn();
+        $("#group-" + id).find(".groupMessage").delay(DELAY_FEEDBACK).fadeOut();
+      }
+    },
+    error: function (err) {
+      console.log(err);
+    },
+  });
+};
+
+var setupDefaultActions = function(id, is_default) {
+  if (is_default) {
+    $("#group-" + id).find("#GroupDefault").html(str_unset_default);
+    $("#group-" + id).find(".is-default-token").attr("title",str_unset_default)
+    $("#group-" + id).find(".is-default-token").removeClass("deactivate");
+    $("#group-" + id).find("#GroupDefault").unbind("click");
+    $("#group-" + id).find("#GroupDefault").on("click", function(){setDefaultGroup(id, false)})
+    $("#group-" + id).find(".is-default-token").on("click", function(){setDefaultGroup(id, false)})
+  } else {
+    $("#group-" + id).find("#GroupDefault").html(str_set_default);
+    $("#group-" + id).find(".is-default-token").attr("title",str_set_default)
+    $("#group-" + id).find(".is-default-token").addClass("deactivate");;
+    $("#group-" + id).find("#GroupDefault").unbind("click");
+    $("#group-" + id).find("#GroupDefault").on("click", function(){setDefaultGroup(id, true)})
+    $("#group-" + id).find(".is-default-token").unbind("click");
+  }
 }
 
 
@@ -287,14 +359,12 @@ $(function () {
     if ($(this).is(":checked")) {
       $(".in-selection-mode").show();
       $(".not-in-selection-mode").hide();
-      $("#group_name").attr("contenteditable", false);
       $(".GroupManagerButtons").removeClass("visible");
     } else {
       $(".in-selection-mode").fadeOut();
       $(".not-in-selection-mode").removeAttr("style");
       $(".Group-checkbox input").attr("checked", false);
       $(".Group-checkbox input[type='checkbox']").trigger("change");
-      $("#group_name").attr("contenteditable", true);
     }
   });
 });
@@ -397,3 +467,185 @@ $('.ConfirmDeleteButton').on("click", function() {
     updateSelectionPanel("NoSelection");
   })
 })
+
+/*-------
+ Manage User Part
+ -------*/
+
+// Initialize the research user bar
+var selectize;
+
+// List of users
+var usersSearch = [];
+
+// Setup the user research bar
+$(function() {
+
+  // initialize the Selectize control
+  $select = $('.AddUserBlock input').selectize({
+    delimiter: ',',
+    persist: false,
+    plugins: ['remove_button']
+  });
+
+  // fetch the instance
+  selectize = $select[0].selectize;
+
+  var idSearch = "";
+  $('.UserSearch input').on("focus", function() {
+    // Import users if it is not already done
+    if (usersSearch.length == 0) {
+      jQuery.ajax({
+        url: "ws.php?format=json&method=pwg.users.getList",
+        type: "POST",
+        data: "",
+        success: function (raw_data) {
+          data = jQuery.parseJSON(raw_data);
+          if (data.stat === "ok") {
+            usersSearch = data.result.users;
+            updateUserSearch();
+            selectize.refreshOptions();
+          }
+        }
+      });
+    } else if (idSearch != $("#UserList").attr("data-group_id")) {
+      updateUserSearch();
+    }
+  });
+
+  // Update User search bar (remove group users in selection)
+  updateUserSearch = function () {
+    idSearch = $("#UserList").attr("data-group_id");
+    selectize.clearOptions();
+    usersSearch.forEach(function(u){
+      isInGroup = false;
+      $('.UsernameBlock').each(function(){
+        if ($(this).data("id")==u.id)
+          isInGroup = true;
+      })
+      if (!isInGroup) {
+        selectize.addOption({value:u.id, text:u.username})
+      }
+    })
+  }
+});
+
+// Display the user manager for a specific group
+var openUserManager = function(grp_id) {
+  jQuery.ajax({
+    url: "ws.php?format=json&method=pwg.users.getList",
+    type: "POST",
+    data: "group_id=" + grp_id,
+    success: function (raw_data) {
+      data = jQuery.parseJSON(raw_data);
+      if (data.stat === "ok") {
+        //Fill with user blocks
+        let users = data.result.users;
+        $(".UsersInGroupList").html('');
+        $(".UserNumberBadge").html(''+users.length);
+        users.forEach(u => {
+          addUserDisplay(u.username, u.id, grp_id);
+        });
+        $('#UserList').fadeIn();
+        //Attribute the group id to the div
+        $("#UserList").attr("data-group_id", grp_id);
+        //Clear the selection
+        selectize.clear();
+      }
+    },
+    error: function (err) {
+      console.log(err);
+    },
+  });
+}
+
+//Add a user block
+var addUserDisplay = function(username, user_id, grp_id) {
+  let userBlock = $('<div class="UsernameBlock" data-id='+user_id+'>'+
+      '<span class="icon-user-1"></span>'+
+      '<p>'+username+'</p>'+
+      '<div class="Tooltip">'+
+        '<span class="icon-cancel"></span>'+
+        '<p class="TooltipText">Dissociate user from this group</p>'+
+      '</div>'+
+      '<div class="UserInfo"><p>User Dissociated</p></div>'+
+    '</div>');
+    userBlock.appendTo(".UsersInGroupList");
+
+    //Setup the delete action
+    userBlock.find(".icon-cancel").on("click", function () {
+    jQuery.ajax({
+      url: "ws.php?format=json&method=pwg.groups.deleteUser",
+      type: "POST",
+      data: "group_id=" + grp_id + "&user_id=" + user_id + "&pwg_token=" + pwg_token,
+      success: function (raw_data) {
+        data = jQuery.parseJSON(raw_data);
+        if (data.stat === "ok") {
+          //Setup User Info
+          userBlock.find(".UserInfo")
+            .css("display", "flex")
+            .addClass("UserInfo-dissociated").removeClass("UserInfo-associated")
+            .hide()
+            .find("p").html(str_user_dissociated);
+          userBlock.find(".UserInfo").fadeIn();
+          userBlock.delay(1000).fadeOut(function(){
+            userBlock.remove();
+            updateUserSearch();
+            //Update member number
+            $(".GroupContainer[data-id="+grp_id+"] .group_number_users")
+            .html(($(".UsernameBlock").length) + " " + str_member_default);
+            $(".UserNumberBadge").html(''+$(".UsernameBlock").length);
+          })
+        }
+      }
+    });
+  })
+  return userBlock;
+} 
+
+// Close pop-up on cross click
+$(".CloseUserList").on("click", function() {$('#UserList').fadeOut();})
+
+// Adding Group Action
+$(".AddUserBlock button").on("click", function () {
+  let grp_id = $("#UserList").attr("data-group_id")
+  let usersString = ""
+  // Get selected ids
+  let ids = selectize.getValue();
+  ids.split(',').forEach(function(id){
+    usersString += "&user_id[]="+id
+  });
+  jQuery.ajax({
+    url: "ws.php?format=json&method=pwg.groups.addUser",
+    type: "POST",
+    data: "group_id=" + grp_id+ usersString + "&pwg_token=" + pwg_token,
+    success: function (raw_data) {
+      data = jQuery.parseJSON(raw_data);
+      ids.split(',').forEach(function(id){
+        // Get the username
+        let username = "undefined";
+        usersSearch.forEach(function(u) {
+          if (u.id == id) {
+            username = u.username;
+          }
+        })
+        let userBlock = addUserDisplay(username, id, grp_id)
+
+        //Setup User Info
+        userBlock.find(".UserInfo")
+          .css("display", "flex")
+          .addClass("UserInfo-associated").removeClass("UserInfo-dissociated")
+          .show()
+          .find("p").html(str_user_associated);
+        userBlock.delay(1000).fadeIn();
+        userBlock.find(".UserInfo").delay(1000).fadeOut();
+
+        updateUserSearch();
+        //Update member number
+        $(".GroupContainer[data-id="+grp_id+"] .group_number_users")
+        .html(($(".UsernameBlock").length) + " " + str_member_default);
+        $(".UserNumberBadge").html(''+$(".UsernameBlock").length);
+      })
+    }
+  }); 
+});
